feat: refactor ActionButtons component to use custom hook for player actions
